# wms_api/api-contracts.yaml
openapi: 3.0.3
info:
  title: WMS API - Warehouse Management System
  description: |
    API контракты для системы управления складом (WMS).
    Включает управление товарами, заказами, размещением, пользователями и логистикой доставки.
  version: 1.0.0
  contact:
    name: WMS Development Team
servers:
  - url: http://localhost:3002/api
    description: Development server (WMS API)
  - url: http://localhost:3001/api
    description: Mock server (wms_site/server)
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: Получить список товаров
      operationId: getProducts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: brand
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsListResponse'

    post:
      tags: [Products]
      summary: Создать товар
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{productId}:
    get:
      tags: [Products]
      summary: Получить товар по ID
      operationId: getProductById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '200':
          description: Данные товара
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Products]
      summary: Обновить товар
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Товар обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

    delete:
      tags: [Products]
      summary: Удалить товар
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '204':
          description: Товар удален
...
  /products/categories:
    get:
      tags: [Products]
      summary: Получить список категорий
      operationId: getCategories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список категорий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

  # ---------------------------------------------------------------------------
  # PLACEMENT
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
...
    ProductIdParam:
      name: productId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # --- Auth ---
...
    ProductStatus:
      type: string
      enum: [pending, processing, completed, rejected]
    Product:
      type: object
      properties:
        id:
          type: integer
        barcode:
          type: string
        name:
          type: string
        brand:
          type: string
        country:
          type: string
        category:
          type: string
        image:
          type: string
        weight:
          type: string
        dimensions:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        location:
          type: string
        packageType:
          type: string
        containerBarcode:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateProductRequest:
      type: object
      required: [barcode, name, category]
      properties:
        barcode:
          type: string
        name:
          type: string
        brand:
          type: string
        country:
          type: string
        category:
          type: string
        image:
          type: string
        weight:
          type: string
        dimensions:
          type: string
        quantity:
          type: integer
        price:
          type: number
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        country:
          type: string
        category:
          type: string
        image:
          type: string
        weight:
          type: string
        dimensions:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        location:
          type: string
        quantity:
          type: integer
        price:
          type: number
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        parentId:
          type: integer
          nullable: true

    # --- Placement ---
...